C# Simple Calculator
A simple calculator for those who are new to the C# language or want to give it a last chance.
Includes a few extra checks than other calculators because we're not kids.
This is my first console project with C#. I hope this helps someone like me who is a beginner in this language. Enjoy :)